research_tool/requirements.txt
python-dotenv
streamlit
unstructured
tiktoken
langchain
langchain-core
langchain-chroma
langchain-groq
langchain-community
langchain-text-splitters
protobuf
langchain-huggingface
sentence-transformers
pypdf
python-docx
docx2txt
pillow
pytesseract
fastapi
uvicorn[standard]
python-multipart